+++ pom.xml
@@ -344,7 +344,6 @@
         <configuration>
           <compilerArgs>
             <arg>-parameters</arg>
-            <arg>-Xlint:-options</arg>
           </compilerArgs>
         </configuration>
       </plugin>
