+++ src/main/java/us/ullberg/startpunkt/crd/BookmarkList.java
@@ -3,14 +3,14 @@ package us.ullberg.startpunkt.crd;
 import io.fabric8.kubernetes.api.model.DefaultKubernetesResourceList;
 
 /**
- * Represents a list of {@link Bookmark} custom resources.
- * Extends {@link DefaultKubernetesResourceList} to provide standard list behavior.
+ * Represents a list of {@link Bookmark} custom resources. Extends {@link
+ * DefaultKubernetesResourceList} to provide standard list behavior.
  */
 public class BookmarkList extends DefaultKubernetesResourceList<Bookmark> {
 
   /**
-   * Creates an empty BookmarkList.
-   * Required to satisfy tools that enforce explicit constructor documentation.
+   * Creates an empty BookmarkList. Required to satisfy tools that enforce explicit constructor
+   * documentation.
    */
   public BookmarkList() {
     super();
