+++ src/main/resources/application.properties
@@ -18,13 +18,11 @@ quarkus.container-image.tag=latest
 
 quarkus.http.test-timeout=60s
 
-quarkus.native.resources.includes=i8n/*.json
 quarkus.swagger-ui.always-include=true
 
 quarkus.native.additional-build-args=--initialize-at-build-time
-quarkus.native.resources.includes=reflection-config.json
 
-quarkus.native.resources.includes=i8n/*.json
+quarkus.native.resources.includes=i8n/*.json,reflection-config.json
 
 quarkus.index-dependency.startpunkt.group-id=us.ullberg
 quarkus.index-dependency.startpunkt.artifact-id=startpunkt
