+++ src/main/resources/application.yaml
@@ -20,7 +20,7 @@ startpunkt:
   # OpenShift integration, read OpenShift Routes
   openshift:
     enabled: true  # If true, enable the reading of OpenShift Route resources
-    onlyAnnotated: true  # Only consider resources with the annotation 'startpunkt.ullberg.us/enable: "true"'
+    onlyAnnotated: false  # Only consider resources with the annotation 'startpunkt.ullberg.us/enable: "true"'
 
   # Hajimari integration, read Hajimari Applications and Bookmarks
   hajimari:
@@ -36,7 +36,7 @@ startpunkt:
   gatewayapi:
     httproute:
       enabled: false  # If true, enable the reading of Hajimari Applications and Bookmarks
-      onlyAnnotated: true  # Only consider resources with the annotation 'startpunkt.ullberg.us/enable: "true"'
+      onlyAnnotated: false  # Only consider resources with the annotation 'startpunkt.ullberg.us/enable: "true"'
 
   # Web ui configuration
   web:
